+++ src/client/confirm_history.rs
@@ -8,6 +8,8 @@ use crate::core::replicon_tick::RepliconTick;
 ///
 /// For efficiency we store only the last received tick and
 /// a bitmask indicating whether the most recent 64 ticks were received.
+///
+/// See also [`EntityReplicated`].
 #[derive(Component)]
 pub struct ConfirmHistory {
     /// Previously confirmed ticks, including the last tick at position 0.
@@ -124,8 +126,17 @@ impl ConfirmHistory {
     }
 }
 
-#[deprecated(note = "use `ConfirmHistory` instead")]
-pub type Confirmed = ConfirmHistory;
+/// Triggered for an entity when it receives updates for a tick.
+///
+/// See also [`ConfirmHistory`].
+#[derive(Debug, Event, Clone, Copy)]
+pub struct EntityReplicated {
+    /// Entity that received an update.
+    pub entity: Entity,
+
+    /// Message tick.
+    pub tick: RepliconTick,
+}
 
 #[cfg(test)]
 mod tests {