+++ src/lib.rs
@@ -623,9 +623,6 @@ pub mod server;
 pub mod test_app;
 
 pub mod prelude {
-    #[allow(deprecated)]
-    pub use super::core::replication::Replication;
-
     pub use super::{
         core::{
             channels::{ChannelKind, RepliconChannel, RepliconChannels},