+++ src/server/replication_messages/mutate_message.rs
@@ -13,10 +13,10 @@ use crate::core::{
 
 /// A message with replicated component mutations.
 ///
-/// Contains change tick, current tick, mutate index and component mutations since
+/// Contains update tick, current tick, mutate index and component mutations since
 /// the last acknowledged tick for each entity.
 ///
-/// Cannot be applied on the client until the change message matching this message's change tick
+/// Cannot be applied on the client until the update message matching this message's update tick
 /// has been applied to the client world.
 /// The message will be manually split into packets up to max size, and each packet will be applied
 /// independently on the client.
@@ -43,7 +43,7 @@ pub(crate) struct MutateMessage {
     /// Components are stored in multiple chunks because some clients may acknowledge mutations,
     /// while others may not.
     ///
-    /// Unlike [`ChangeMessage`](super::change_message::ChangeMessage), we serialize the number
+    /// Unlike [`UpdateMessage`](super::update_message::UpdateMessage), we serialize the number
     /// of chunk bytes instead of the number of components. This is because, during deserialization,
     /// some entities may be skipped if they have already been updated (as mutations are sent until
     /// the client acknowledges them).
@@ -59,7 +59,6 @@ pub(crate) struct MutateMessage {
     /// Intermediate buffer with mutate index, message size and a range for [`Self::mutations`].
     ///
     /// We split messages first in order to know their count in advance.
-    /// We plan to include it in the message in the future.
     messages: Vec<(u16, usize, Range<usize>)>,
 }
 
@@ -124,19 +123,24 @@ impl MutateMessage {
         client: &mut ReplicatedClient,
         client_buffers: &mut ClientBuffers,
         serialized: &SerializedData,
+        track_mutate_messages: bool,
         server_tick: Range<usize>,
         tick: Tick,
         timestamp: Duration,
     ) -> bincode::Result<usize> {
         debug_assert_eq!(self.entities.len(), self.mutations.len());
 
-        let mut change_tick = Cursor::new([0; mem::size_of::<RepliconTick>()]);
-        bincode::serialize_into(&mut change_tick, &client.change_tick())?;
-        let ticks_size = change_tick.get_ref().len() + server_tick.len();
+        const MAX_COUNT_SIZE: usize = mem::size_of::<usize>() + 1;
+        let mut update_tick = Cursor::new([0; mem::size_of::<RepliconTick>()]);
+        bincode::serialize_into(&mut update_tick, &client.update_tick())?;
+        let mut metadata_size = update_tick.get_ref().len() + server_tick.len();
+        if track_mutate_messages {
+            metadata_size += MAX_COUNT_SIZE;
+        }
 
         let (mut mutate_index, mut entities) =
             client.register_mutate_message(client_buffers, tick, timestamp);
-        let mut header_size = ticks_size + mutate_index.required_space();
+        let mut header_size = metadata_size + mutate_index.required_space();
         let mut body_size = 0;
         let mut mutations_range = Range::<usize>::default();
         for (entity, mutations) in self.entities.iter().zip(&self.mutations) {
@@ -158,7 +162,7 @@ impl MutateMessage {
                 mutations_range.start = mutations_range.end;
                 (mutate_index, entities) =
                     client.register_mutate_message(client_buffers, tick, timestamp);
-                header_size = ticks_size + mutate_index.required_space(); // Recalculate since the mutate index changed.
+                header_size = metadata_size + mutate_index.required_space(); // Recalculate since the mutate index changed.
                 body_size = 0;
             }
 
@@ -166,8 +170,9 @@ impl MutateMessage {
             mutations_range.end += 1;
             body_size += mutations_size;
         }
-        if !mutations_range.is_empty() {
+        if !mutations_range.is_empty() || track_mutate_messages {
             // When the loop ends, pack all leftovers into a message.
+            // Or create an empty message if tracking mutate messages is enabled.
             self.messages.push((
                 mutate_index,
                 body_size + header_size,
@@ -176,11 +181,18 @@ impl MutateMessage {
         }
 
         let messages_count = self.messages.len();
-        for (mutate_index, message_size, mutations_range) in self.messages.drain(..) {
+        for (mutate_index, mut message_size, mutations_range) in self.messages.drain(..) {
+            if track_mutate_messages {
+                // Update message counter size based on actual value.
+                message_size -= MAX_COUNT_SIZE - messages_count.required_space();
+            }
             let mut message = Vec::with_capacity(message_size);
 
-            message.extend_from_slice(change_tick.get_ref());
+            message.extend_from_slice(update_tick.get_ref());
             message.extend_from_slice(&serialized[server_tick.clone()]);
+            if track_mutate_messages {
+                message.write_varint(messages_count)?;
+            }
             message.write_varint(mutate_index)?;
             for mutations in &self.mutations[mutations_range.clone()] {
                 message.extend_from_slice(&serialized[mutations.entity.clone()]);